move function showMessage and showAdminMessage into new file
