+++ myBitboxMining/webnode-mining/modelMgrs/TokenManager.js
@@ -3,7 +3,7 @@ const TokenModel = require('../models/Token');
 const SequelizeConfig = require('./SequelizeConfig');
 
 // import const
-const showMessage = require('../global/FileHelper').showMessage;
+const showMessage = require('../global/ResourceHelper').showMessage;
 
 const Sequelize = SequelizeConfig.getSequelizeModule();
