+++ myBitboxMining/webnode-mining/private/reactjs/customer/components/account/forgotPassword.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 // import const
-const showMessage = require('../../../../../global/FileHelper').showMessage;
+const showMessage = require('../../../../../global/ResourceHelper').showMessage;
 
 export default class ForgotPassword extends Component {
     constructor (props) {
