+++ myBitboxMining/webnode-mining/private/reactjs/customer/components/wallet/listWalletComponent.js
@@ -4,7 +4,7 @@ import Modal from 'react-modal';
 import WalletItemContainer from '../../containers/wallet/walletItemContainer';
 
 // import const
-const showMessage = require('../../../../../global/FileHelper').showMessage;
+const showMessage = require('../../../../../global/ResourceHelper').showMessage;
 
 // import const
 import { API_URL } from '../../const/variable';
